+++ tests/Datasets/LabeledTest.php
@@ -103,10 +103,10 @@ class LabeledTest extends TestCase
         $this->assertEquals(self::SAMPLES, $this->dataset->samples());
     }
 
-    public function test_get_row()
+    public function test_get_sample()
     {
-        $this->assertEquals(self::SAMPLES[2], $this->dataset->row(2));
-        $this->assertEquals(self::SAMPLES[5], $this->dataset->row(5));
+        $this->assertEquals(self::SAMPLES[2], $this->dataset->sample(2));
+        $this->assertEquals(self::SAMPLES[5], $this->dataset->sample(5));
     }
 
     public function test_num_rows()
@@ -442,7 +442,7 @@ class LabeledTest extends TestCase
     {
         $kernel = new Gower(9.);
 
-        [$left, $right] = $this->dataset->spatialPartition($this->dataset->row(0), $this->dataset->row(3), $kernel);
+        [$left, $right] = $this->dataset->spatialPartition($this->dataset->sample(0), $this->dataset->sample(3), $kernel);
 
         $this->assertInstanceOf(Labeled::class, $left);
         $this->assertInstanceOf(Labeled::class, $right);
@@ -482,7 +482,7 @@ class LabeledTest extends TestCase
 
         $this->assertCount(count(self::SAMPLES) + 1, $merged);
 
-        $this->assertEquals(['nice', 'furry', 'friendly'], $merged->row(0));
+        $this->assertEquals(['nice', 'furry', 'friendly'], $merged->sample(0));
         $this->assertEquals('not monster', $merged->label(6));
     }
 
@@ -496,7 +496,7 @@ class LabeledTest extends TestCase
 
         $this->assertCount(count(self::SAMPLES) + 1, $merged);
 
-        $this->assertEquals(['nice', 'furry', 'friendly'], $merged->row(6));
+        $this->assertEquals(['nice', 'furry', 'friendly'], $merged->sample(6));
         $this->assertEquals('not monster', $merged->label(6));
     }