+++ tests/Graph/Trees/ITreeTest.php
@@ -55,7 +55,7 @@ class ITreeTest extends TestCase
 
         $this->assertGreaterThan(5, $this->tree->height());
 
-        $sample = $this->generator->generate(1)->row(0);
+        $sample = $this->generator->generate(1)->sample(0);
 
         $node = $this->tree->search($sample);