+++ tests/Transformers/GaussianRandomProjectorTest.php
@@ -36,7 +36,7 @@ class GaussianRandomProjectorTest extends TestCase
 
     public function test_fit_transform()
     {
-        $this->assertCount(10, $this->generator->generate(1)->row(0));
+        $this->assertCount(10, $this->generator->generate(1)->sample(0));
 
         $this->transformer->fit($this->generator->generate(30));
 
@@ -44,7 +44,7 @@ class GaussianRandomProjectorTest extends TestCase
 
         $sample = $this->generator->generate(1)
             ->apply($this->transformer)
-            ->row(0);
+            ->sample(0);
 
         $this->assertCount(5, $sample);
     }