+++ tests/Transformers/IntervalDiscretizerTest.php
@@ -39,7 +39,7 @@ class IntervalDiscretizerTest extends TestCase
 
         $sample = $this->generator->generate(1)
             ->apply($this->transformer)
-            ->row(0);
+            ->sample(0);
 
         $this->assertCount(4, $sample);