+++ tests/Transformers/MinMaxNormalizerTest.php
@@ -41,7 +41,7 @@ class MinMaxNormalizerTest extends TestCase
 
         $sample = $this->generator->generate(1)
             ->apply($this->transformer)
-            ->row(0);
+            ->sample(0);
 
         $this->assertCount(3, $sample);