+++ tests/Transformers/RobustStandardizerTest.php
@@ -37,7 +37,7 @@ class RobustStandardizerTest extends TestCase
 
         $sample = $this->generator->generate(1)
             ->apply($this->transformer)
-            ->row(0);
+            ->sample(0);
 
         $this->assertCount(3, $sample);