+++ tests/Transformers/VarianceThresholdFilterTest.php
@@ -37,7 +37,7 @@ class VarianceThresholdFilterTest extends TestCase
 
         $sample = $this->generator->generate(1)
             ->apply($this->transformer)
-            ->row(0);
+            ->sample(0);
 
         $this->assertCount(2, $sample);