+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - Added training and inference sections to the docs
     - Decision tree rules method now outputs a string
     - Added drop row and column methods to dataset interface
+    - Dataset row() method is now sample()
 
 - 0.0.16-beta
     - Radius Neighbors allows user-definable anomaly class